+++ demos/category-pan.html
@@ -312,7 +312,7 @@ firstDayArr.forEach(obj => {
   })
 })
 
-chart.interact('catPan', {
+chart.interact('category-pan', {
   mode: 'x',
   onProcess: () => {
   },
