+++ src/interaction/index.js
@@ -1,8 +1,8 @@
 module.exports = {
   Interaction: require('./base'),
   PieSelect: require('./pie-select'),
-  Pan: require('./pan'),
-  Zoom: require('./zoom'),
-  CatPan: require('./category-pan'),
-  CatPinch: require('./category-pinch')
+  LinearPan: require('./linear-pan'),
+  LinearPinch: require('./linear-pinch'),
+  CategoryPan: require('./category-pan'),
+  CategoryPinch: require('./category-pinch')
 };
